requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
_ipython_minor_entry_point=8.7.0=hb6b4a82_0
aom=3.5.0=h7ea286d_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
brotli=1.0.9=h1a8c8d9_8
brotli-bin=1.0.9=h1a8c8d9_8
brotlipy=0.7.0=py310h8e9501a_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.12.7=h4653dfc_0
cairo=1.16.0=h73a0509_1014
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h2399d43_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
comm=0.1.2=pyhd8ed1ab_0
contourpy=1.0.6=py310h2887b22_0
cryptography=38.0.4=py310hfc83b78_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.4=py310h0f1eb42_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=hb7217d7_0
ffmpeg=5.1.2=gpl_hf318d42_105
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.1=h82840c6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=py310h8e9501a_1
freetype=2.12.1=hd633e50_1
gettext=0.21.1=h0186832_0
gmp=6.2.1=h9f76cd9_0
gnutls=3.7.8=h9f1a10d_0
graphite2=1.3.13=h9f76cd9_1001
harfbuzz=6.0.0=hddbc195_0
hdf5=1.12.2=nompi_h33dac16_100
icu=70.1=h6b3803e_0
idna=3.4=pyhd8ed1ab_0
ipykernel=6.19.2=pyh736e0ef_0
ipython=8.7.0=pyhd1c38e8_0
jasper=2.0.33=hba35424_0
jedi=0.18.2=pyhd8ed1ab_0
jpeg=9e=he4db4b2_2
jupyter_client=7.4.8=pyhd8ed1ab_0
jupyter_core=5.1.0=py310hbe9552e_0
kiwisolver=1.4.4=py310h2887b22_1
krb5=1.20.1=h69eda48_0
lame=3.100=h1a8c8d9_1003
lcms2=2.14=h8193b64_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=16_osxarm64_openblas
libbrotlicommon=1.0.9=h1a8c8d9_8
libbrotlidec=1.0.9=h1a8c8d9_8
libbrotlienc=1.0.9=h1a8c8d9_8
libcblas=3.9.0=16_osxarm64_openblas
libcurl=7.86.0=h9049daf_2
libcxx=14.0.6=h2692d47_0
libdeflate=1.14=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=11_3_0_hd922786_27
libgfortran5=11.3.0=hdaf2cc0_27
libglib=2.74.1=h4646484_1
libiconv=1.17=he4db4b2_0
libidn2=2.3.4=h1a8c8d9_0
liblapack=3.9.0=16_osxarm64_openblas
liblapacke=3.9.0=16_osxarm64_openblas
libnghttp2=1.47.0=h519802c_1
libopenblas=0.3.21=openmp_hc731615_3
libopencv=4.6.0=py310h57eab40_8
libpng=1.6.39=h76d750c_0
libprotobuf=3.21.12=hb5ab8b9_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=h7a5bd25_3
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.4.0=heb92581_5
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.10.3=h87b0503_0
libzlib=1.2.13=h03a7124_4
llvm-openmp=15.0.6=h7cfbb63_0
matplotlib=3.6.2=py310hb6292c7_0
matplotlib-base=3.6.2=py310h78c5c2f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.8.1=h63371fa_1
ninja=1.11.0=hf86a087_0
numpy=1.23.5=py310h5d7c261_0
opencv=4.6.0=py310hb6292c7_8
openh264=2.3.1=hb7217d7_1
openjpeg=2.5.0=h5d4e404_1
openssl=3.0.7=h03a7124_1
p11-kit=0.24.1=h29577a5_0
packaging=22.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.40=hb34f9b4_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py310h9337a76_3
pip=22.3.1=pyhd8ed1ab_0
pixman=0.40.0=h27ca646_0
platformdirs=2.6.0=pyhd8ed1ab_0
prompt-toolkit=3.0.36=pyha770c72_0
psutil=5.9.4=py310h8e9501a_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-opencv=4.6.0=py310h69fb684_8
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.8=h3ba56d0_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=1.13.0=cpu_py310h7410233_0
pyzmq=24.0.1=py310hc407298_1
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_1
setuptools=65.6.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=1.4.1=h7ea286d_0
tk=8.6.12=he1e0b03_0
torchvision=0.14.0=cpu_py310h7df52bc_0
tornado=6.2=py310h8e9501a_1
traitlets=5.7.1=pyhd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
unicodedata2=15.0.0=py310h8e9501a_0
urllib3=1.26.13=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.38.4=pyhd8ed1ab_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
zeromq=4.3.4=hbdafb3b_1
zlib=1.2.13=h03a7124_4
zstd=1.5.2=h8128057_4
